+++ src/ObjectTreeWalker/ObjectAccessor.cs
@@ -36,8 +36,15 @@ internal class ObjectAccessor
 		_objectType = objectType;
 		foreach (var propertyInfo in objectType.GetProperties(BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Public))
 		{
-			_getPropertyMethods.Add(propertyInfo.Name, CreateGetPropertyFunc(propertyInfo));
-			_setPropertyMethods.Add(propertyInfo.Name, CreateSetPropertyFunc(propertyInfo));
+			if (propertyInfo.GetMethod != null)
+			{
+				_getPropertyMethods.Add(propertyInfo.Name, CreateGetPropertyFunc(propertyInfo));
+			}
+
+			if (propertyInfo.SetMethod != null)
+			{
+				_setPropertyMethods.Add(propertyInfo.Name, CreateSetPropertyFunc(propertyInfo));
+			}
 		}
 
 		foreach (var fieldInfo in objectType.GetFields(BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Public))
@@ -157,7 +164,7 @@ internal class ObjectAccessor
 		if (fieldInfo.FieldType.IsValueType)
 		{
 			emitter.LoadArgument(1);
-			emitter.BranchIfTrue("after");
+			emitter.BranchIfTrue("after"); // if null it will not branch!
 
 			var getDefaultConcrete =
 				GetDefaultCache.GetOrAdd(
@@ -216,7 +223,7 @@ internal class ObjectAccessor
 			emitter.CastClass(_objectType);
 		}
 
-		emitter.CallVirtual(propertyInfo.GetMethod);
+		emitter.Call(propertyInfo.GetMethod);
 
 		if (propertyInfo.PropertyType.IsValueType)
 		{
@@ -274,7 +281,7 @@ internal class ObjectAccessor
 			emitter.CastClass(propertyInfo.PropertyType);
 		}
 
-		emitter.CallVirtual(propertyInfo.SetMethod);
+		emitter.Call(propertyInfo.SetMethod);
 		emitter.Return();
 
 		return emitter.CreateDelegate();