+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.1.0] - 2022-09-17
+### :white_check_mark: Tests
+- [`6b07a38`](https://github.com/myarichuk/ObjectTreeWalker/commit/6b07a38b23228f6517b3547e5450477aa7268d07) - ensure ObjectAccessor tests work on both struct and class objects *(commit by [@myarichuk](https://github.com/myarichuk))*
+
+
 ## [v2.1.0.0] - 2022-09-14
 ### :sparkles: New Features
 - [`0e85a6c`](https://github.com/myarichuk/Library.Template/commit/0e85a6c62eee9c66c38785e062ce8337169c982e) - automatically fill release notes when generating nuspec *(commit by [@myarichuk](https://github.com/myarichuk))*
@@ -43,3 +48,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v2.0.0.0]: https://github.com/myarichuk/Library.Template/compare/v1.0.23.0...v2.0.0.0
 [v2.0.1.0]: https://github.com/myarichuk/Library.Template/compare/v2.0.0.0...v2.0.1.0
 [v2.1.0.0]: https://github.com/myarichuk/Library.Template/compare/v2.0.1.0...v2.1.0.0
+[v0.2.1.0]: https://github.com/myarichuk/ObjectTreeWalker/compare/v0.2.0.0...v0.2.1.0